moved it out of the api folder and created a new route
