+++ routes/index.js
@@ -2,9 +2,11 @@ const router = require('express').Router();
 const apiRoutes = require('./api');
 const homeRoutes = require('./homeRoutes.js');
 const dashboardRoutes = require('./dashboardRoutes.js');
+const filteredRestaurants = require('./filteredRouters.js');
 
 router.use('/api', apiRoutes);
 router.use('/', homeRoutes);
 router.use('/dashboard', dashboardRoutes);
+router.use('/filter', filteredRestaurants);
 
 module.exports = router;
